Add kustomize
Kustomize was used in the earlier helm3 implementation. It
is necessary that it is used in an identical way on apps
that are associated with helm3 charts and were migrated
from apps.project.cattle.io. Otherwise, the absence of
comparable kustomize logic may cause manifests lacking
existing matchLabel selectors, which are immutable, to be
applied resulting in errors. Kustomize logic similar to
what was implemented before has been added for migrated
apps.