Only install kustomize for amd64 and arm64
Prior, if the arch was arm, a download of kustomize for arm would
be attempted. Kustomize does not currently support arm so the
download would fail. Now, a download will only be attempted for
amd64 and arm64. This means that kustomize will not be available
if a shell pod is on an arm node.